+++ packages/core/styles/src/packages/atomic/icon/nimbus-icon.css.ts
@@ -49,7 +49,7 @@ const iconColorProperties = {
   "neutral-textHigh": colorProperties["neutral-textHigh"],
 
   // This color will be overwritten by the Icon SVG gradient processor
-  "ai-interactive": "currentColor",
+  "ai-generative": "currentColor",
 };
 
 const properties = {